AllGames/learn_multiplication/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-cors-headers==4.7.0
django-webpack-loader==3.1.1
djangorestframework==3.15.2
gunicorn==20.1.0
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2024.2
